python_eda/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38hca72f7f_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38hecd8cb5_0
biopython=1.79=py38hed1de0f_2
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py38h4277f33_1
ca-certificates=2022.6.15=h033912b_0
certifi=2022.6.15=py38h50d1736_0
cffi=1.15.0=py38hca72f7f_0
colorama=0.4.5=pyhd8ed1ab_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=h18a8e69_0
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.2.10=h1c7c35f_0
freetype=2.10.4=h4cff582_1
gettext=0.19.8.1=haf92f58_1004
glib=2.58.3=py38h338c551_1004
icu=64.2=h6de7cb9_1
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.9.1=py38hecd8cb5_0
ipython=8.4.0=py38hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
jsonschema=4.4.0=py38hecd8cb5_0
jupyter=1.0.0=py38_7
jupyter_client=7.2.2=py38hecd8cb5_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py38he9d5cce_0
libblas=3.9.0=15_osx64_openblas
libcblas=3.9.0=15_osx64_openblas
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.2.1=h0a44026_1007
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=15_osx64_openblas
libllvm11=11.1.0=hd011deb_3
libopenblas=0.3.20=openmp_hb3cd9ec_0
libpng=1.6.37=h7cec526_2
libsodium=1.0.18=h1de35cc_0
libzlib=1.2.12=hfe4f2af_2
llvm-openmp=14.0.4=ha654fa7_0
llvmlite=0.38.0=py38h8346a28_0
markupsafe=2.1.1=py38hca72f7f_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h1300a51_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h1de35cc_1001
nbclient=0.5.13=py38hecd8cb5_0
nbconvert=6.4.4=py38hecd8cb5_0
nbformat=5.3.0=py38hecd8cb5_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.5=py38hecd8cb5_0
notebook=6.4.12=py38hecd8cb5_0
numba=0.55.1=py38hae1ba45_0
numexpr=2.8.0=py38ha53d530_1
numpy=1.20.3=py38h5cb586d_2
openssl=1.1.1q=hfe4f2af_0
opentsne=0.3.11=py38h1de35cc_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38he9d5cce_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.4=py38hecd8cb5_0
plotly=5.8.2=py_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pynndescent=0.5.7=pyh6c4a22f_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.18.0=py38hca72f7f_0
python=3.8.0=h359304d_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2022.1=pyhd8ed1ab_0
pyzmq=23.2.0=py38he9d5cce_0
qt=5.9.7=h8cf7e54_3
qtconsole=5.3.1=py38hecd8cb5_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=7.0=h1de35cc_5
scikit-learn=1.0.2=py38h37f3bb3_0
scipy=1.8.0=py38hd329d04_1
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py38hecd8cb5_0
sip=4.19.8=py38h0a44026_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.33.0=hffcf06c_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=py38hc1426ef_0
tbb=2021.5.0=h940c156_0
tenacity=8.0.1=py38hecd8cb5_0
terminado=0.13.1=py38hecd8cb5_0
testpath=0.6.0=py38hecd8cb5_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5d9f67b_0
tornado=6.1=py38hed1de0f_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
umap-learn=0.5.3=py38h50d1736_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38hecd8cb5_0
xz=5.2.5=hca72f7f_1
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py38hecd8cb5_0
zlib=1.2.12=h4dc903c_2
